move Qt stuff into its own file
